% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indep_pval.R
\name{get_indep_pval}
\alias{get_indep_pval}
\title{Generate p-value for the independent test}
\usage{
get_indep_pval(X, Y, pve)
}
\arguments{
\item{X}{A matrix of functions}

\item{Y}{A matrix of functions}

\item{pve}{proportion varaince explained; a value between 0-1}
}
\description{
Internal function
}
\keyword{internal}
