% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paired_pvalue.R
\name{get_paired_pvalue}
\alias{get_paired_pvalue}
\title{Generate p-value for the paired test}
\usage{
get_paired_pvalue(X, Y, no_perm = 1000)
}
\arguments{
\item{X}{A matrix of functions}

\item{Y}{A matrix of functions}

\item{n_perm}{Number of permutations used to calculate p-value for paired test}
}
\description{
Internal function
}
\keyword{internal}
