% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paired_statistic.R
\name{get_paired_statistic}
\alias{get_paired_statistic}
\title{Estimate the test statistic for the paired data}
\usage{
get_paired_statistic(X, Y)
}
\arguments{
\item{X}{A matrix of functions}

\item{Y}{A matrix of functions}
}
\description{
Internal function
}
\keyword{internal}
